move source, event, eventhandler, predicate, reconcile packages under pkg/
