+++ pkg/eventhandler/enqueue_owner.go
@@ -19,8 +19,8 @@ package eventhandler
 import (
 	"fmt"
 
-	"github.com/kubernetes-sigs/controller-runtime/pkg/controller/event"
-	"github.com/kubernetes-sigs/controller-runtime/pkg/controller/reconcile"
+	"github.com/kubernetes-sigs/controller-runtime/pkg/event"
+	"github.com/kubernetes-sigs/controller-runtime/pkg/reconcile"
 	"github.com/kubernetes-sigs/controller-runtime/pkg/runtime/inject"
 	logf "github.com/kubernetes-sigs/controller-runtime/pkg/runtime/log"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
