+++ pkg/predicate/doc.go
